+++ Source/Lib/Directory.Build.props
@@ -2,7 +2,7 @@
 	<PropertyGroup>
 		<TargetFrameworks>netstandard2.0;netstandard2.1;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
 		<LangVersion>10</LangVersion>
-		<Version>5.5.0-Beta1</Version>
+		<Version>5.5.0</Version>
 		<AssemblyVersion>5.5.0.0</AssemblyVersion>
 		<FileVersion>5.5.0.0</FileVersion>
