+++ Source/Tests/Fluxor.UnitTests/FeatureTests/SupportFiles/IsolatedTests.cs
@@ -1,0 +1,6 @@
+﻿namespace Fluxor.UnitTests.FeatureTests.SupportFiles
+{
+	public class IsolatedTests : Middleware
+	{
+	}
+}
