+++ Source/Tests/Fluxor.UnitTests/FeatureTests/SupportFiles/TestState.cs
@@ -1,0 +1,12 @@
+﻿namespace Fluxor.UnitTests.FeatureTests.SupportFiles
+{
+	public class TestState
+	{
+		public int SomeValue { get; }
+
+		public TestState(int someValue)
+		{
+			SomeValue = someValue;
+		}
+	}
+}
